step 3 수정
