+++ CodeStarterCamp_Week1/step 2.swift
@@ -1,0 +1,23 @@
+//
+//  step 2.swift
+//  CodeStarterCamp_Week1
+//
+//  Created by 조기열 on 2022/06/29.
+//
+
+import Foundation
+
+let iceCream = "***********"
+let stick = "    | |"
+
+func drawIceCream() {
+    for _ in 1...8 {
+        print("\(iceCream)")
+    }
+}
+
+func drawStick() {
+    for _ in 1...4 {
+        print("\(stick)")
+    }
+}
